refactor: PresignedUrlResponse 패키지 이
